All linker and compiler issues resolved. MILESTONE ACHIEVED!!
EUSJGALA176:CC-ONF-driver edeedhu$ $MAKE all
mkdir /Users/edeedhu/hub/forked/CC-ONF-driver/obj
gcc -I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/Users/edeedhu/hub/forked/CC-ONF-driver/include -Wall -Wextra -g  -c /Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_pollthr_mgr.c -o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_pollthr_mgr.o
gcc -I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/Users/edeedhu/hub/forked/CC-ONF-driver/include -Wall -Wextra -g  -c /Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_lib.c -o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_lib.o
/Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_lib.c: In function ‘cc_of_lib_abort’:
/Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_lib.c:76: warning: control reaches end of non-void function
/Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_lib.c: In function ‘cc_of_strerror’:
/Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_lib.c:85: warning: return discards qualifiers from pointer target type
gcc -I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/Users/edeedhu/hub/forked/CC-ONF-driver/include -Wall -Wextra -g  -c /Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_tcp_conn.c -o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_tcp_conn.o
/Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_tcp_conn.c:26: warning: initialization from incompatible pointer type
/Users/edeedhu/hub/forked/CC-ONF-driver/include/cc_of_lib.h:21: warning: ‘cc_of_errtable’ defined but not used
gcc -I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/Users/edeedhu/hub/forked/CC-ONF-driver/include -Wall -Wextra -g  -c /Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_of_util.c -o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_util.o
/Users/edeedhu/hub/forked/CC-ONF-driver/include/cc_of_lib.h:21: warning: ‘cc_of_errtable’ defined but not used
gcc -I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/Users/edeedhu/hub/forked/CC-ONF-driver/include -Wall -Wextra -g  -c /Users/edeedhu/hub/forked/CC-ONF-driver/src/cc_udp_conn.c -o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_udp_conn.o
/Users/edeedhu/hub/forked/CC-ONF-driver/include/cc_of_lib.h:21: warning: ‘cc_of_errtable’ defined but not used
	gcc -lglib-2.0 -shared /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_pollthr_mgr.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_lib.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_tcp_conn.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_util.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_udp_conn.o -o libccof.so
gcc -lglib-2.0 -shared /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_pollthr_mgr.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_lib.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_tcp_conn.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_of_util.o /Users/edeedhu/hub/forked/CC-ONF-driver/obj/cc_udp_conn.o -o libccof.so
EUSJGALA176:CC-ONF-driver edeedhu$
EUSJGALA176:CC-ONF-driver edeedhu$
EUSJGALA176:CC-ONF-driver edeedhu$
EUSJGALA176:CC-ONF-driver edeedhu$ ls -l libccof.so
-rwxr-xr-x  1 edeedhu  2110378014  45584 Aug 13 18:10 libccof.so
EUSJGALA176:CC-ONF-driver edeedhu$
EUSJGALA176:CC-ONF-driver edeedhu$
EUSJGALA176:CC-ONF-driver edeedhu$ ls -l obj/
total 192
-rw-r--r--  1 edeedhu  2110378014  16376 Aug 13 18:10 cc_of_lib.o
-rw-r--r--  1 edeedhu  2110378014  21796 Aug 13 18:10 cc_of_util.o
-rw-r--r--  1 edeedhu  2110378014  28520 Aug 13 18:10 cc_pollthr_mgr.o
-rw-r--r--  1 edeedhu  2110378014  12808 Aug 13 18:10 cc_tcp_conn.o
-rw-r--r--  1 edeedhu  2110378014   9228 Aug 13 18:10 cc_udp_conn.o

+++ include/cc_net_conn.h
@@ -94,12 +94,12 @@ typedef struct net_svcs_ {
     int (*open_clientfd)(char *ipaddr, int port);
     int (*open_serverfd)(char *ipaddr, int port);
     int (*accept_conn)(int listenfd, struct sockaddr  *clientaddr, 
-	               int *addrlen);
+	               uint32_t *addrlen);
     int (*close_conn)(int *sockfd);
-    int (*read_data)(int sockfd, void *buf, size_t len, int flags,
-	             struct sockaddr *src_addr, socklen_t *addrlen);
-    int (*write_data)(int sockfd, const void *buf, size_t len, int flags,
-	              const struct sockaddr *dest_addr, socklen_t addrlen);
+    ssize_t (*read_data)(int sockfd, void *buf, size_t len, int flags,
+                         struct sockaddr *src_addr, socklen_t *addrlen);
+    ssize_t (*write_data)(int sockfd, const void *buf, size_t len, int flags,
+                          const struct sockaddr *dest_addr, socklen_t addrlen);
 } net_svcs_t;
    
 #endif //CC_NET_CONN_H
